Fixed the unit tests after the refactor.
