+++ tests/unit/writers/qbwriter/test_qbwriter.py
@@ -331,629 +331,6 @@ def test_virtual_columns_generated_for_multi_meas_obs_val():
     assert "cube-name.csv#unit/some-unit" == virt_unit["valueUrl"]
 
 
-def test_serialise_new_attribute_values():
-    """
-    When new attribute values are serialised, a list of new metadata resources should be returned.
-    """
-
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [2, 2, 2],
-            "New Attribute": ["Pending", "Final", "In Review"],
-            "Existing Attribute": ["D", "E", "F"],
-        }
-    )
-
-    metadata = CatalogMetadata("Some Dataset")
-    columns = [
-        QbColumn(
-            "Existing Dimension",
-            ExistingQbDimension("https://example.org/dimensions/existing_dimension"),
-        ),
-        QbColumn(
-            "Value",
-            QbObservationValue(
-                ExistingQbMeasure("http://example.org/existing/measure"),
-                ExistingQbUnit("http://example.org/some/existing/unit"),
-            ),
-        ),
-        QbColumn(
-            "New Attribute",
-            NewQbAttribute.from_data("New Attribute", data["New Attribute"]),
-        ),
-        QbColumn(
-            "Existing Attribute",
-            ExistingQbAttribute(
-                "http://example.org/some/existing/attribute",
-                new_attribute_values=[
-                    NewQbAttributeValue(
-                        "D",
-                        description="real value",
-                        parent_attribute_value_uri="http://parent-uri",
-                    ),
-                    NewQbAttributeValue("E", source_uri="http://source-uri"),
-                    NewQbAttributeValue("F"),
-                ],
-            ),
-        ),
-    ]
-
-    cube = Cube(metadata, data, columns)
-
-    qbwriter = QbWriter(cube)
-    list_of_new_attribute_values = qbwriter._get_new_attribute_value_resources()
-
-    map_label_to_expected_config = {
-        "Pending": {
-            "uri": "some-dataset.csv#attribute/new-attribute/pending",
-            "description": None,
-            "parent_attribute_value_uri": None,
-            "source_uri": None,
-        },
-        "Final": {
-            "uri": "some-dataset.csv#attribute/new-attribute/final",
-            "description": None,
-            "parent_attribute_value_uri": None,
-            "source_uri": None,
-        },
-        "In Review": {
-            "uri": "some-dataset.csv#attribute/new-attribute/in-review",
-            "description": None,
-            "parent_attribute_value_uri": None,
-            "source_uri": None,
-        },
-        "D": {
-            "uri": "some-dataset.csv#attribute/existing-attribute/d",
-            "description": "real value",
-            "parent_attribute_value_uri": "http://parent-uri",
-            "source_uri": None,
-        },
-        "E": {
-            "uri": "some-dataset.csv#attribute/existing-attribute/e",
-            "description": None,
-            "parent_attribute_value_uri": None,
-            "source_uri": "http://source-uri",
-        },
-        "F": {
-            "uri": "some-dataset.csv#attribute/existing-attribute/f",
-            "description": None,
-            "parent_attribute_value_uri": None,
-            "source_uri": None,
-        },
-    }
-
-    for (label, expected_config) in map_label_to_expected_config.items():
-        new_attribute_value = first(
-            list_of_new_attribute_values, lambda x: x.label == label
-        )
-        assert new_attribute_value is not None
-        assert new_attribute_value.uri_str == expected_config["uri"]
-        assert new_attribute_value.comment == expected_config["description"]
-
-        assert (
-            expected_config["parent_attribute_value_uri"] is None
-            and new_attribute_value.parent_attribute_value_uri is None
-        ) or str(new_attribute_value.parent_attribute_value_uri.uri) == expected_config[
-            "parent_attribute_value_uri"
-        ]
-
-        assert (
-            expected_config["source_uri"] is None
-            and new_attribute_value.source_uri is None
-        ) or str(new_attribute_value.source_uri.uri) == expected_config["source_uri"]
-
-
-def test_serialise_unit():
-    """
-    When new units are serialised, a list of new metadata resources should be returned.
-    """
-
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [2, 2, 2],
-            "Units": ["Percent", "People", "People"],
-        }
-    )
-
-    metadata = CatalogMetadata("Some Dataset")
-    columns = [
-        QbColumn(
-            "Existing Dimension",
-            ExistingQbDimension("https://example.org/dimensions/existing_dimension"),
-        ),
-        QbColumn(
-            "Value",
-            QbObservationValue(
-                ExistingQbMeasure("http://example.org/existing/measure")
-            ),
-        ),
-        QbColumn(
-            "Units",
-            QbMultiUnits(
-                [
-                    NewQbUnit(
-                        "Percent",
-                        description="unit",
-                        base_unit=ExistingQbUnit("http://parent-uri"),
-                    ),
-                    NewQbUnit(
-                        "People",
-                        source_uri="http://source-uri",
-                        si_base_unit_conversion_multiplier=25.1364568,
-                        qudt_quantity_kind_uri="http://some-quantity-kind-family",
-                    ),
-                ],
-            ),
-        ),
-    ]
-
-    cube = Cube(metadata, data, columns)
-
-    qbwriter = QbWriter(cube)
-    list_of_new_unit_resources = qbwriter._get_new_unit_resources()
-
-    map_label_to_expected_config = {
-        "Percent": {
-            "uri": "some-dataset.csv#unit/percent",
-            "description": "unit",
-            "parent_unit_uri": "http://parent-uri",
-            "source_uri": None,
-            "qudt_quantity_kind_family": None,
-            "qudt_conversion_multiplier": None,
-        },
-        "People": {
-            "uri": "some-dataset.csv#unit/people",
-            "description": None,
-            "parent_unit_uri": None,
-            "source_uri": "http://source-uri",
-            "qudt_quantity_kind_family": "http://some-quantity-kind-family",
-            "qudt_conversion_multiplier": 25.1364568,
-        },
-    }
-    for (label, expected_config) in map_label_to_expected_config.items():
-        new_attribute_value = first(
-            list_of_new_unit_resources, lambda x: x.label == label
-        )
-        assert new_attribute_value is not None
-        assert new_attribute_value.uri_str == expected_config["uri"]
-        assert new_attribute_value.comment == expected_config["description"]
-        assert (
-            expected_config["source_uri"] is None
-            and new_attribute_value.source_uri is None
-        ) or str(new_attribute_value.source_uri.uri) == expected_config["source_uri"]
-        assert (
-            expected_config["parent_unit_uri"] is None
-            and new_attribute_value.base_unit_uri is None
-        ) or str(new_attribute_value.base_unit_uri.uri) == expected_config[
-            "parent_unit_uri"
-        ]
-
-        assert (
-            expected_config["qudt_conversion_multiplier"] is None
-            and new_attribute_value.qudt_conversion_multiplier is None
-        ) or new_attribute_value.qudt_conversion_multiplier == expected_config[
-            "qudt_conversion_multiplier"
-        ]
-        assert (
-            expected_config["qudt_quantity_kind_family"] is None
-            and new_attribute_value.has_qudt_quantity_kind is None
-        ) or str(new_attribute_value.has_qudt_quantity_kind.uri) == expected_config[
-            "qudt_quantity_kind_family"
-        ]
-
-
-def test_arbitrary_rdf_serialisation_existing_attribute():
-    """
-    Test that when arbitrary RDF is specified against an Existing Attribute, it is serialised correctly.
-    """
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [1, 2, 3],
-            "Existing Attribute": ["Provisional", "Final", "Provisional"],
-        }
-    )
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "Existing Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension"
-                ),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-            QbColumn(
-                "Existing Attribute",
-                ExistingQbAttribute(
-                    "http://some-existing-attribute-uri",
-                    arbitrary_rdf=[
-                        TripleFragment(RDFS.label, "Existing Attribute Component")
-                    ],
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#component/existing-attribute"),
-        RDFS.label,
-        Literal("Existing Attribute Component"),
-    ) in graph
-
-
-def test_arbitrary_rdf_serialisation_new_attribute():
-    """
-    Test that when arbitrary RDF is specified against a New Attribute, it is serialised correctly.
-    """
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [1, 2, 3],
-            "New Attribute": ["Provisional", "Final", "Provisional"],
-        }
-    )
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "Existing Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension"
-                ),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-            QbColumn(
-                "Existing Attribute",
-                NewQbAttribute.from_data(
-                    "New Attribute",
-                    data["New Attribute"],
-                    arbitrary_rdf=[
-                        TripleFragment(RDFS.label, "New Attribute Property"),
-                        TripleFragment(
-                            RDFS.label,
-                            "New Attribute Component",
-                            RdfSerialisationHint.Component,
-                        ),
-                    ],
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#attribute/new-attribute"),
-        RDFS.label,
-        Literal("New Attribute Property"),
-    ) in graph
-
-    assert (
-        URIRef("some-dataset.csv#component/new-attribute"),
-        RDFS.label,
-        Literal("New Attribute Component"),
-    ) in graph
-
-
-def test_arbitrary_rdf_serialisation_existing_dimension():
-    """
-    Test that when arbitrary RDF is specified against an Existing Dimension, it is serialised correctly.
-    """
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [1, 2, 3],
-            "Existing Attribute": ["Provisional", "Final", "Provisional"],
-        }
-    )
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "Existing Dimension",
-                ExistingQbDimension(
-                    "https://example.org/dimensions/existing_dimension",
-                    arbitrary_rdf=[
-                        TripleFragment(RDFS.label, "Existing Dimension Component")
-                    ],
-                ),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#component/existing-dimension"),
-        RDFS.label,
-        Literal("Existing Dimension Component"),
-    ) in graph
-
-
-def test_arbitrary_rdf_serialisation_new_dimension():
-    """
-    Test that when arbitrary RDF is specified against a new dimension, it is serialised correctly.
-    """
-    data = pd.DataFrame({"New Dimension": ["A", "B", "C"], "Value": [1, 2, 3]})
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "New Dimension",
-                NewQbDimension.from_data(
-                    "Some Dimension",
-                    data["New Dimension"],
-                    arbitrary_rdf=[
-                        TripleFragment(RDFS.label, "New Dimension Property"),
-                        TripleFragment(
-                            RDFS.label,
-                            "New Dimension Component",
-                            RdfSerialisationHint.Component,
-                        ),
-                    ],
-                ),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#dimension/some-dimension"),
-        RDFS.label,
-        Literal("New Dimension Property"),
-    ) in graph
-
-    assert (
-        URIRef("some-dataset.csv#component/some-dimension"),
-        RDFS.label,
-        Literal("New Dimension Component"),
-    ) in graph
-
-
-def test_arbitrary_rdf_serialisation_new_dimension_with_cube_uri_style_WithoutFileExtensions():
-    """
-    Test that when arbitrary RDF is specified against a new dimension, it is serialised correctly.
-    """
-    data = pd.DataFrame({"New Dimension": ["A", "B", "C"], "Value": [1, 2, 3]})
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "New Dimension",
-                NewQbDimension.from_data(
-                    "Some Dimension",
-                    data["New Dimension"],
-                    arbitrary_rdf=[
-                        TripleFragment(RDFS.label, "New Dimension Property"),
-                        TripleFragment(
-                            RDFS.label,
-                            "New Dimension Component",
-                            RdfSerialisationHint.Component,
-                        ),
-                    ],
-                ),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"), NewQbUnit("Some Unit")
-                ),
-            ),
-        ],
-        uri_style=URIStyle.WithoutFileExtensions,
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset#dimension/some-dimension"),
-        RDFS.label,
-        Literal("New Dimension Property"),
-    ) in graph
-
-    assert (
-        URIRef("some-dataset#component/some-dimension"),
-        RDFS.label,
-        Literal("New Dimension Component"),
-    ) in graph
-
-
-def test_arbitrary_rdf_serialisation_existing_dimension():
-    """
-    Test that when arbitrary RDF is specified against an Existing Dimension, it is serialised correctly.
-    """
-    data = pd.DataFrame(
-        {
-            "Existing Dimension": ["A", "B", "C"],
-            "Value": [1, 2, 3],
-            "Existing Attribute": ["Provisional", "Final", "Provisional"],
-        }
-    )
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "Existing Dimension",
-                NewQbDimension.from_data(
-                    "Existing Dimension", data["Existing Dimension"]
-                ),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    ExistingQbMeasure(
-                        "http://some/uri/existing-measure-uri",
-                        arbitrary_rdf=[
-                            TripleFragment(RDFS.label, "Existing Measure Component")
-                        ],
-                    ),
-                    NewQbUnit("Some Unit"),
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#component/existing-measure-uri"),
-        RDFS.label,
-        Literal("Existing Measure Component"),
-    ) in graph
-
-
-def test_arbitrary_rdf_serialisation_new_measure():
-    """
-    Test that when arbitrary RDF is specified against a new measure, it is serialised correctly.
-    """
-    data = pd.DataFrame({"New Dimension": ["A", "B", "C"], "Value": [1, 2, 3]})
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "New Dimension",
-                NewQbDimension.from_data("Some Dimension", data["New Dimension"]),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure(
-                        "Some Measure",
-                        arbitrary_rdf=[
-                            TripleFragment(RDFS.label, "New Measure Property"),
-                            TripleFragment(
-                                RDFS.label,
-                                "New Measure Component",
-                                RdfSerialisationHint.Component,
-                            ),
-                        ],
-                    ),
-                    NewQbUnit("Some Unit"),
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#measure/some-measure"),
-        RDFS.label,
-        Literal("New Measure Property"),
-    ) in graph
-
-    assert (
-        URIRef("some-dataset.csv#component/some-measure"),
-        RDFS.label,
-        Literal("New Measure Component"),
-    ) in graph
-
-
-def test_qb_order_of_components():
-    """
-    Test that when components are created they have a qb:order value.
-    """
-    data = pd.DataFrame({"New Dimension": ["A", "B", "C"], "Value": [1, 2, 3]})
-
-    cube = Cube(
-        CatalogMetadata("Some Dataset"),
-        data,
-        [
-            QbColumn(
-                "New Dimension",
-                NewQbDimension.from_data("Some Dimension", data["New Dimension"]),
-            ),
-            QbColumn(
-                "Value",
-                QbObservationValue(
-                    NewQbMeasure("Some Measure"),
-                    NewQbUnit("Some Unit"),
-                ),
-            ),
-        ],
-    )
-
-    qb_writer = QbWriter(cube)
-    dataset = qb_writer._generate_qb_dataset_dsd_definitions()
-    graph = dataset.to_graph(Graph())
-
-    assert (
-        URIRef("some-dataset.csv#component/some-dimension"),
-        rdf.QB.order,
-        Literal(1),
-    ) in graph
-
-    assert (URIRef("some-dataset.csv#component/measure-type"), rdf.QB.order, Literal(2))
-
-    assert (
-        URIRef("some-dataset.csv#component/unit"),
-        rdf.QB.order,
-        Literal(3),
-    ) in graph
-
-    assert (
-        URIRef("some-dataset.csv#component/some-measure"),
-        rdf.QB.order,
-        Literal(4),
-    ) in graph
-
-
 def test_output_integer_obs_val_with_missing_values():
     """
     Due to the way that pandas represents missing values (as NaN), instead of interpreting an integer data column as
